+++ .github/workflows/main.yml
@@ -1,0 +1,32 @@
+name: CI
+
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    branches:
+      - master
+
+jobs:
+
+  test:
+    name: Testing on ${{matrix.os}} and Node ${{matrix.node_version}}
+    runs-on: ${{matrix.os}}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest]
+        node_version: [10, 12, 13]
+    steps:
+      - name: Checkout Master
+        uses: actions/checkout@v1
+      - name: Use Node ${{matrix.node_version}}
+        uses: actions/setup-node@master
+        with:
+          version: ${{ matrix.node_version }}
+      - name: Install Dependencies using Yarn
+        run: yarn install
+      - name: Build
+        run: yarn build
+      - name: Test
+        run: yarn test
