+++ .gitignore
@@ -15,3 +15,5 @@ results
 
 npm-debug.log
 dist/
+yarn.lock
+package-lock.json
